Added navBar and globalBackground
Created a NavBar that navigates to Home, Search, Lists and Accounts. This is applied to all pages. Functionality will needed to be added so it is not applied to LandingPage. This can be done after AuthViewModel is created.

Also created a custom view modifier for global background color. This can be applied easily to a ZStack on every page. Easy to update color that will apply to all pages.