Added tracker - work in progress
Fleshed out most of the Book Tracker, added a form, fields, buttons. Created a ViewModel and a Model. There are a couple of bugs/adjustments to make for the form, including a nil end date, formatting issues in the form at the bottom, allowing the user to select out of the decimalpad, adding user validation to ensure they cannot enter a value greater than the book length, and making sure the bottom nav bar appear on this view. The Model also needs to be linked to the BookModel and UserAuthentication. Then the data needs to be stored in the cloud.

Also added small elements tot he HomePageView but need to add the BookPreviews to it.

Made comment on the custom button to have the button adjust size based on different text lengths.